fixed some code related to check available version
Version checking within the method is basically treated as if it has already been done.

However, if the version that the calling struct can handle is different from the version that the method can handle, version checking is performed on the method.
